Add files via upload
Unlike simple #objectclassification such as distinguishing cats from dogs, Flower classification is a challenging task due to the wide range of flower species, which have a similar shape, appearance or surrounding objects such as leaves and grass. The 𝐓𝐫𝐚𝐧𝐬𝐟𝐞𝐫 𝐥𝐞𝐚𝐫𝐧𝐢𝐧𝐠 with fine-tuning reaches 87% 𝘢𝘤𝘤𝘶𝘳𝘢𝘤𝘺 on testing data, which is higher than custom #CNNmodels and all the traditional methods.